requirements.txt
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
pycparser==2.22
PySocks==1.7.1
robotframework==7.2.2
robotframework-pythonlibcore==4.4.1
robotframework-seleniumlibrary==6.7.1
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
